% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rli-package.R
\docType{package}
\name{rli-package}
\alias{rli}
\alias{rli-package}
\title{rli: Access the LinkedIn API}
\description{
A consistent interface to the LinkedIn API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/rli}
  \item \url{https://jonthegeek.github.io/rli/}
  \item Report bugs at \url{https://github.com/jonthegeek/rli/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
